LayoutTests/accessibility/aria-grid-with-aria-owns-rows-expected.txt

Foo
Bar
This tests that an ARIA table can use aria-owns for its cells.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS row1.childAtIndex(0).isEqual(table.cellForColumnAndRow(0, 0)) is true
PASS row1.childAtIndex(0).isEqual(accessibilityController.accessibleElementById('row1-cell-1')) is true
PASS cell1.parentElement().isEqual(row1) is true
PASS row1.parentElement().isEqual(table) is true
PASS successfullyParsed is true

TEST COMPLETE
